Move Assignment interface to types folder
