+++ ui/src/Assignments/AssignmentForm.tsx
@@ -21,7 +21,6 @@ import AssignmentClient from 'Assignments/AssignmentClient';
 import SelectWithNoCreateOption, {MetadataMultiSelectProps} from 'ModalFormComponents/SelectWithNoCreateOption';
 import {Option} from 'Types/Option';
 import {ProductPlaceholderPair} from './CreateAssignmentRequest';
-import {Assignment} from './Assignment';
 import FormButton from 'ModalFormComponents/FormButton';
 import SelectWithCreateOption, {MetadataReactSelectProps} from 'ModalFormComponents/SelectWithCreateOption';
 import {useRecoilValue, useSetRecoilState} from 'recoil';
@@ -35,6 +34,7 @@ import {Product} from '../Types/Product';
 import {Person} from '../Types/Person';
 
 import './AssignmentForm.scss';
+import {Assignment} from '../Types/Assignment';
 
 interface Props {
     initiallySelectedProduct: Product;
