+++ ui/src/Assignments/History/AssignmentHistory.test.tsx
@@ -23,14 +23,16 @@ import {AssignmentHistory} from './AssignmentHistory';
 import ProductClient from '../../Products/ProductClient';
 import moment, {now} from 'moment';
 import {fireEvent} from '@testing-library/dom';
-import {Assignment} from '../Assignment';
+import {Assignment} from '../../Types/Assignment';
 
 jest.mock('Products/ProductClient');
 jest.mock('Assignments/AssignmentClient');
 
 describe('Assignment History', () => {
     const daysBetweenStartAndToday = (assignment: Assignment): number  => {
-        return Math.floor(moment.duration(moment(now()).startOf('day').diff(moment(assignment.startDate).startOf('day'))).asDays());
+        return Math.floor(moment.duration(
+            moment(now()).startOf('day').diff(moment(assignment.startDate).startOf('day'))
+        ).asDays());
     };
 
     async function clickLabel(renderResult: RenderResult): Promise<void> {
