+++ ui/src/DragAndDrop/DragAndDrop.tsx
@@ -23,10 +23,10 @@ import {CurrentSpaceState} from '../State/CurrentSpaceState';
 import AssignmentClient from '../Assignments/AssignmentClient';
 import {ViewingDateState} from '../State/ViewingDateState';
 import moment from 'moment';
-import {Assignment} from '../Assignments/Assignment';
 import {ProductPlaceholderPair} from '../Assignments/CreateAssignmentRequest';
 import useFetchProducts from '../Hooks/useFetchProducts/useFetchProducts';
 import {Product} from '../Types/Product';
+import {Assignment} from '../Types/Assignment';
 
 type Props = {};
