+++ ui/src/Utils/TestData.ts
@@ -15,7 +15,6 @@
  * limitations under the License.
  */
 
-import {Assignment} from 'Assignments/Assignment';
 import {LocationTag, RoleTag, Tag} from 'Types/Tags';
 import {Space} from 'Types/Space';
 import {UserSpaceMapping} from 'Types/UserSpaceMapping';
@@ -23,6 +22,7 @@ import {LocalStorageFilters} from 'SortingAndFiltering/FilterLibraries';
 import {Color} from 'Types/Color';
 import {Product} from 'Types/Product';
 import {Person} from 'Types/Person';
+import {Assignment} from '../Types/Assignment';
 
 const originDateString = '2019-01-01';
